fix: useCoupons hook 내부에서 쿠폰 계산 순수 함수 분리 1.  generateCouponCode 2. validateCoupon 3. canApplyCoupon 4. calculateDiscount
