+++ src/basic/domain/coupon/index.ts
@@ -10,3 +10,6 @@ export * from "./hooks";
 
 // Constants
 export * from "./constants";
+
+// Utils
+export * from "./utils";
